bug 1159406 - Move v1-specific classes and tests
To prepare to serve multiple API versions at once:
* Create a new folder for v1-specific code
* Move v1-specific classes (parsers, renderers) into new v1 folder
* Split API classes like viewsets into common and v1-specific classes
* Rearrange tests between general and v1-specific tests
