+++ webplatformcompat/routers.py
@@ -12,15 +12,6 @@ from rest_framework.routers import DefaultRouter
 from rest_framework.urlpatterns import format_suffix_patterns
 from rest_framework.views import APIView
 
-from .viewsets import (
-    BrowserViewSet, FeatureViewSet, MaturityViewSet, SectionViewSet,
-    SpecificationViewSet, SupportViewSet, VersionViewSet,
-    HistoricalBrowserViewSet, HistoricalFeatureViewSet,
-    HistoricalMaturityViewSet, HistoricalSectionViewSet,
-    HistoricalSpecificationViewSet, HistoricalSupportViewSet,
-    HistoricalVersionViewSet,
-    ChangesetViewSet, UserViewSet, ViewFeaturesViewSet)
-
 
 class GroupedRouter(DefaultRouter):
     """Router with grouped API root and slash redirects."""
@@ -102,37 +93,3 @@ class GroupedRouter(DefaultRouter):
                     pattern_name=u.name, permanent=False, query_string=True)
                 redirect_urls.append(url(pattern, view))
         return urls + redirect_urls
-
-
-router = GroupedRouter(trailing_slash=False, version='v1')
-
-router.register(r'browsers', BrowserViewSet, group='resources')
-router.register(r'versions', VersionViewSet, group='resources')
-router.register(r'features', FeatureViewSet, group='resources')
-router.register(r'supports', SupportViewSet, group='resources')
-router.register(r'specifications', SpecificationViewSet, group='resources')
-router.register(r'maturities', MaturityViewSet, group='resources')
-router.register(r'sections', SectionViewSet, group='resources')
-
-router.register(r'changesets', ChangesetViewSet, group='change_control')
-router.register(r'users', UserViewSet, group='change_control')
-
-router.register(
-    r'historical_browsers', HistoricalBrowserViewSet, group='history')
-router.register(
-    r'historical_versions', HistoricalVersionViewSet, group='history')
-router.register(
-    r'historical_features', HistoricalFeatureViewSet, group='history')
-router.register(
-    r'historical_supports', HistoricalSupportViewSet, group='history')
-router.register(
-    r'historical_specifications', HistoricalSpecificationViewSet,
-    group='history')
-router.register(
-    r'historical_maturities', HistoricalMaturityViewSet, group='history')
-router.register(
-    r'historical_sections', HistoricalSectionViewSet, group='history')
-
-router.register(
-    r'view_features', ViewFeaturesViewSet, base_name='viewfeatures',
-    group='views')
