+++ webplatformcompat/tests/v1/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for the v1 API."""
