+++ webplatformcompat/urls.py
@@ -3,7 +3,7 @@ from django.conf.urls import include, patterns, url
 from django.views.generic import TemplateView
 from django.views.generic.base import RedirectView
 from mdn.urls import mdn_urlpatterns
-from webplatformcompat.routers import router
+from webplatformcompat.v1.routers import router as v1_router
 
 from .views import ViewFeature
 
@@ -23,7 +23,7 @@ webplatformcompat_urlpatterns = patterns(
         namespace='rest_framework')),
     url(r'^api/$', RedirectView.as_view(url='/api/v1/', permanent=False),
         name='api_root'),
-    url(r'^api/v1/', include(router.urls, namespace='v1')),
+    url(r'^api/v1/', include(v1_router.urls, namespace='v1')),
     url(r'^importer$', RedirectView.as_view(
         url='/importer/', permanent=False)),
     url(r'^importer/', include(mdn_urlpatterns)),
