+++ webplatformcompat/v1/__init__.py
@@ -1,0 +1,4 @@
+"""Implementation of the v1 API.
+
+The v1 API supports JSON API Release Candidate 1 (RC1)
+"""
